# src/swagger/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Fortress API",
		"version": "1.0.0",
		"description": "Fortress is a social media platform that allows men to connect, share experiences, and support each other in a safe and empowering environment. The platform is designed to foster positive masculinity, mental health awareness, and community engagement through various features such as posts, comments, and user profiles. This is achieved through anonimity, robust moderation, and a focus on constructive interactions."
	},
	"paths": {
...
		"/comment/post/{postId}": {
			"post": {
				"tags": ["Comment"],
				"summary": "Create comment on post",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Comment created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"get": {
				"tags": ["Comment"],
				"summary": "Get comments for a post",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "parentCommentId",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Comments list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentsListResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/like": {
			"post": {
				"tags": ["Like"],
				"summary": "Like a target (post or comment)",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLikeDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Liked",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LikeResponse"
								}
							}
						}
					},
					"400": {
						"description": "Already liked or invalid",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"get": {
				"tags": ["Like"],
				"summary": "Get likes",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "targetType",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["post", "comment"]
						}
					},
					{
						"name": "targetId",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Likes list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LikesListResponse"
								}
							}
						}
					}
				}
			}
		},
		"/like/{targetType}/{targetId}": {
			"delete": {
				"tags": ["Like"],
				"summary": "Unlike target",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "targetType",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"enum": ["post", "comment"]
						}
					},
					{
						"name": "targetId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Unliked",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Like not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/like/user/{userId}": {
			"get": {
				"tags": ["Like"],
				"summary": "Get likes by user",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "User likes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LikesListResponse"
								}
							}
						}
					}
				}
			}
		},
		"/like/check/{targetType}/{targetId}": {
			"get": {
				"tags": ["Like"],
				"summary": "Check if current user liked the target",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "targetType",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"enum": ["post", "comment"]
						}
					},
					{
						"name": "targetId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Check result",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"liked": {
											"type": "boolean"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/notification": {
			"post": {
				"tags": ["Notification"],
				"summary": "Create notification",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateNotificationDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Notification created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"get": {
				"tags": ["Notification"],
				"summary": "Get notifications",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "unreadOnly",
						"in": "query",
						"schema": {
							"type": "boolean"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Notifications list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationsListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/notification/unread-count": {
			"get": {
				"tags": ["Notification"],
				"summary": "Get unread count",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Unread count",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"unreadCount": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/notification/{id}/read": {
			"patch": {
				"tags": ["Notification"],
				"summary": "Mark notification as read",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Marked as read",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/notification/read-all": {
			"patch": {
				"tags": ["Notification"],
				"summary": "Mark all notifications as read",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "All marked",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"/notification/{id}": {
				"delete": {
					"tags": ["Notification"],
					"summary": "Delete notification",
					"security": [
						{
							"bearerAuth": []
						}
					],
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"responses": {
						"200": {
							"description": "Deleted",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/GenericMessageResponse"
									}
								}
							}
						},
						"404": {
							"description": "Not found",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"401": {
							"description": "Unauthorized"
						}
					}
				}
			}
		},
...
		"/post": {
			"get": {
				"tags": ["Post"],
				"summary": "Generate feed",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					},
					{
						"name": "userId",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "hashtag",
						"in": "query",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Feed",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Post"
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			},
			"post": {
				"tags": ["Post"],
				"summary": "Create a new post",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Post created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/post/{id}": {
			"get": {
				"tags": ["Post"],
				"summary": "Get post by id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Post",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Post"],
				"summary": "Update post",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePostDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Post updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Post"],
				"summary": "Delete post",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericMessageResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/user/{id}": {
			"get": {
				"tags": ["User"],
				"summary": "Get user by id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "User found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"403": {
						"description": "Private profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"User": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"profile": {
						"$ref": "#/components/schemas/UserProfile"
					},
					"counts": {
						"$ref": "#/components/schemas/UserCounts"
					},
					"isVerified": {
						"type": "boolean"
					},
					"isPrivate": {
						"type": "boolean"
					},
					"lastLoginAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"UserProfile": {
				"type": "object",
				"properties": {
					"fullName": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": ["male", "female", "other", "prefer_not_to_say"]
					}
				}
			},
			"UserCounts": {
				"type": "object",
				"properties": {
					"posts": {
						"type": "integer"
					},
					"followers": {
						"type": "integer"
					},
					"following": {
						"type": "integer"
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
			"ValidationErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
			"CreateCommentDto": {
				"type": "object",
				"required": ["text"],
				"properties": {
					"text": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string"
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
...
			"Comment": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"postId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userAvatar": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"parentCommentId": {
						"type": "string"
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"CommentsListResponse": {
				"type": "object",
				"properties": {
					"comments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Comment"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
			"CommentResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Comment"
					}
				}
			},
			"GenericMessageResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					}
				}
			},
...
			"Post": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"caption": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userAvatar": {
						"type": "string"
					},
					"media": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/MediaDto"
						}
					},
					"location": {
						"$ref": "#/components/schemas/LocationDto"
					},
					"hashtags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"counts": {
						"$ref": "#/components/schemas/PostCounts"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"CreateLikeDto": {
				"type": "object",
				"required": ["targetType", "targetId"],
				"properties": {
					"targetType": {
						"type": "string",
						"enum": ["post", "comment"]
					},
					"targetId": {
						"type": "string"
					}
				}
			},
			"Like": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"targetType": {
						"type": "string"
					},
					"targetId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"LikeResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Like"
					}
				}
			},
			"LikesListResponse": {
				"type": "object",
				"properties": {
					"likes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Like"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
			"CoordinatesDto": {
				"type": "object",
				"properties": {
					"latitude": {
						"type": "number"
					},
					"longitude": {
						"type": "number"
					}
				}
			},
			"LocationDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"coordinates": {
						"$ref": "#/components/schemas/CoordinatesDto"
					}
				}
			},
			"MediaDto": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"enum": ["image", "video"]
					},
					"url": {
						"type": "string"
					},
					"thumbnailUrl": {
						"type": "string"
					},
					"width": {
						"type": "number"
					},
					"height": {
						"type": "number"
					},
					"duration": {
						"type": "number"
					}
				}
			},
			"CreatePostDto": {
				"type": "object",
				"required": ["media"],
				"properties": {
					"media": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/MediaDto"
						}
					},
					"caption": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/LocationDto"
					},
					"hashtags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"isCommentsEnabled": {
						"type": "boolean"
					},
					"isLikesVisible": {
						"type": "boolean"
					}
				}
			},
			"UpdatePostDto": {
				"type": "object",
				"properties": {
					"caption": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/LocationDto"
					},
					"hashtags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"mentions": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"isCommentsEnabled": {
						"type": "boolean"
					},
					"isLikesVisible": {
						"type": "boolean"
					}
				}
			},
			"PostCounts": {
				"type": "object",
				"properties": {
					"likes": {
						"type": "integer"
					},
					"comments": {
						"type": "integer"
					}
				}
			},
			"PostResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/Post"
					}
				}
			},
...
			"UserResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/User"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
...
	"security": [{ "bearerAuth": [] }]
}